+++ app/src/main/res/layout/activity_you_tube_detail.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".YouTubeDetailListActivity">
+    tools:context=".YouTubeDetailActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
@@ -17,12 +17,12 @@
             android:backgroundTint="@android:color/white">
 
             <androidx.appcompat.widget.Toolbar
-                android:id="@+id/youtube_detail_list_toolbar"
+                android:id="@+id/youtube_detail_toolbar"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content">
 
                 <ImageView
-                    android:id="@+id/youtube_detail_list_back_btn"
+                    android:id="@+id/youtube_detail_back_btn"
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
                     android:src="@drawable/ic_back_black_24dp"
@@ -31,7 +31,7 @@
                     android:onClick="clickBackBtn"/>
 
                 <TextView
-                    android:id="@+id/youtube_detail_list_top_title"
+                    android:id="@+id/youtube_detail_top_title"
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
                     android:text="@string/youtube_detail_list_top_title"
@@ -44,27 +44,15 @@
         </com.google.android.material.appbar.AppBarLayout>
 
         <RelativeLayout
-            android:id="@+id/youtube_detail_list_relative"
+            android:id="@+id/youtube_detail_relative"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
             android:padding="8dp">
 
-            <ImageView
-                android:id="@+id/youtube_detail_list_search_iv"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:src="@drawable/ic_search_black_24dp"
-                android:layout_alignParentRight="true"
-                android:layout_marginTop="16dp"
-                android:layout_marginRight="8dp"
-                android:clickable="true"
-                android:onClick="clickIv"/>
-
             <EditText
-                android:id="@+id/youtube_detail_list_search_et"
+                android:id="@+id/youtube_detail_search_et"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
-                android:layout_toLeftOf="@id/youtube_detail_list_search_iv"
                 android:background="@drawable/chat_et_bg"
                 android:layout_marginTop="8dp"
                 android:layout_marginRight="8dp"
@@ -74,10 +62,10 @@
                 android:padding="8dp"/>
 
             <Button
-                android:id="@+id/youtube_detail_list_link_youtube_see_btn"
+                android:id="@+id/youtube_detail_link_youtube_see_btn"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
-                android:layout_below="@id/youtube_detail_list_search_et"
+                android:layout_below="@id/youtube_detail_search_et"
                 android:layout_marginTop="8dp"
                 android:background="@drawable/mymenu_btn_selector"
                 android:textAllCaps="false"
@@ -86,24 +74,21 @@
 
 
             <View
-                android:id="@+id/youtube_detail_list_line"
+                android:id="@+id/youtube_detail_line"
                 android:layout_width="match_parent"
                 android:layout_height="1dp"
                 android:background="@android:color/darker_gray"
-                android:layout_below="@id/youtube_detail_list_link_youtube_see_btn"
+                android:layout_below="@id/youtube_detail_link_youtube_see_btn"
                 android:layout_marginBottom="8dp"
                 android:layout_marginTop="16dp"/>
 
-            <!--     들어가야 할 내용
-                   1. -->
-            <androidx.recyclerview.widget.RecyclerView
-                android:id="@+id/youtube_detail_list_detail_youtube_info_recycler"
+            <fragment
+                android:id="@+id/youtube_detail_youtube_view_frag"
                 android:layout_width="match_parent"
-                android:layout_height="match_parent"
-                android:layout_below="@id/youtube_detail_list_line"
-                android:layout_marginTop="8dp"
-                app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
-                android:orientation="vertical"/>
+                android:layout_height="wrap_content"
+                android:name="com.google.android.youtube.player.YouTubePlayerFragment"
+                android:layout_centerVertical="true"
+                android:layout_marginTop="8dp" />
 
 
         </RelativeLayout>
